test: split test_remyx19.py into per-module files
Follows the existing tests/api/ convention of one file per API module:
- tests/api/test_experiments.py  — experiments API tests
- tests/api/test_projects.py     — projects API tests
- tests/api/test_recommendations.py — extended with the repo-sourced
  flow + create/update repo-field tests, alongside the existing
  interests coverage (same pattern as how recommendations and interests
  already co-live in this file)

Ticket-named test files aren't informative about what's tested; module-
named files make it obvious where to add new cases later.
